% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSNACScall.R
\name{clusterCellsWithSNPdata}
\alias{clusterCellsWithSNPdata}
\title{Cluster cells based on the SNP data}
\usage{
clusterCellsWithSNPdata(
  snacsObj,
  clustMethod = c("hclust", "skmean"),
  outputFormat = c("", "pdf", "png")
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{clustMethod}{Character. Clustering method for ranking SNPs. Options are "hclust" and "skmean". Default is "hclust"}

\item{outputFormat}{Character. Output file type. Default is "" which outputs to the standard output}
}
\value{
A SNACSList object
}
\description{
Cluster cells, where the number of clusters is the number of constituent samples, using the mutation data with the selected SNPs. Generate heatmap of the mutation data.
}
