% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depthAltMat}
\alias{depthAltMat}
\title{Example alternate depth data}
\format{
\subsection{\code{depthAltMat}}{

A matrix with 22 rows and 739 columns:
}
}
\usage{
depthAltMat
}
\description{
The alternate depth data should be a numeric matrix with SNPs in the rows and cells in the columns. The total depth value should be a continuous numeric measure.
}
\keyword{datasets}
