% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depthTotalMat}
\alias{depthTotalMat}
\title{Example total depth data}
\format{
\subsection{\code{depthTotalMat}}{

A matrix with 22 rows and 739 columns:
}
}
\usage{
depthTotalMat
}
\description{
The total depth data should be a numeric matrix with SNPs in the rows and cells in the columns. The total depth value should be a continuous numeric measure.
}
\keyword{datasets}
