% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashCall.R
\name{generateHashDensityPlot}
\alias{generateHashDensityPlot}
\title{Make hash calls.}
\usage{
generateHashDensityPlot(
  snacsObj,
  prob_ecdf = 0.95,
  probBackgndPeak = 0.6,
  outputFormat = c("", "pdf", "png")
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{prob_ecdf}{.}

\item{probBackgndPeak}{.}

\item{outputFormat}{Character. Output file type. Default is "" which outputs to the standard output}
}
\description{
Make hash call based on mutation and hash data.
}
