% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestSNPs.R
\name{getBestSNPs.internal}
\alias{getBestSNPs.internal}
\title{Select SNPs that best separates the hashes.}
\usage{
getBestSNPs.internal(
  snacsObj,
  cell_anno_var,
  clustMethod = c("hclust", "skmean"),
  pvSnpThres = 10^-320,
  outputFormat = c("", "pdf", "png")
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{cell_anno_var}{Optional character vector. Cell variables for which color bars are to be added to the heatmap. Default is NULL}

\item{clustMethod}{Character. Clustering method for selecting best SNPs. Options are "hclust" and "skmean". Default is "hclust"}

\item{pvSnpThres}{P-value cutoff to select best SNPs}

\item{outputFormat}{Character. Output file type. Default is "" which outputs to the standard output}
}
\value{
A SNACSList object
}
\description{
Cluster data to select SNPs that best separate the hashes. Generate heatmap of the best SNPs.
}
