% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSNACScall.R
\name{getBestSNPs}
\alias{getBestSNPs}
\title{Select SNPs that best separate the hashes}
\usage{
getBestSNPs(
  snacsObj,
  numSNP = 3,
  backgndThreshold = 0.95,
  cellProportionAboveBackgnd = 0.5,
  cellProportionBelowBackgndMode = 0.6,
  cellProportionForModeDetection = 0.75,
  outputFormat = c("", "pdf", "png")
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{numSNP}{Integer. Number of pairs of SNPs (high in one sample / low in second sample and vice versa) that best separate sample-pairs. There will be s times n(n-1) number of SNPs where s is the number of SNPs and n is the number of samples. Default is 3}

\item{backgndThreshold}{Numeric. Threshold of the background antibody distribution of a hash above which the antibody will be considered to be expressed in a cell. Default is 0.95. Range is 0-1}

\item{cellProportionAboveBackgnd}{Numeric. Proportion of cells in a cluster that have to have signal above the background of a hash, determined by "backgndThreshold" parameter, for the cluster to be assigned to the hash. Default is 0.5. Range is 0-1}

\item{cellProportionBelowBackgndMode}{Numeric. Maximum proportion of cells which can be below the mode of the estimated hash background distribution. Default is 0.6. Range is 0-1}

\item{cellProportionForModeDetection}{Numeric. Proportion of cells used to estimate mode of the background distribution. Used only if "cellProportionBelowBackgndMode" threshold is not met; otherwise, all cells are used. Default is 0.75. Range is 0-1}

\item{outputFormat}{Character. Output file type. Default is "" which outputs to the standard output}
}
\value{
A SNACSList object
}
\description{
Group cells into constituent samples based on hash antibody data. Select the best SNPs that separate the cell groups using the mutation data.
}
