% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSNACScall.R
\name{getHashModeInfo}
\alias{getHashModeInfo}
\title{Get mode information of density function}
\usage{
getHashModeInfo(mm, bw = "nrd0", verbose = FALSE)
}
\arguments{
\item{mm}{Numeric. Hash antibody data}

\item{bw}{Character. Smoothing bandwidth parameter passed to density function}

\item{verbose}{Logical. Prints information when running the method. Default is FALSE}
}
\value{
Mode and adjustment values
}
\description{
Get mode information of hash antibody distribution for background detection.
}
