% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSNACScall.R
\name{getModes}
\alias{getModes}
\title{Get mode of density function}
\usage{
getModes(x, bw, adjust, signifi, from, to)
}
\arguments{
\item{x}{.}

\item{bw}{.}

\item{adjust}{.}

\item{signifi}{.}

\item{from}{.}

\item{to}{.}
}
\value{
A vector of modes
}
\description{
Get two mode density function of hash antibody data for background detection.
}
