% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestSNPs.R
\name{getRankedSNPs.internal}
\alias{getRankedSNPs.internal}
\title{Rank SNPs in the order that best separates the hashes}
\usage{
getRankedSNPs.internal(
  snacsObj,
  cell_anno_var,
  clustMethod = c("hclust", "skmean"),
  outputFormat = c("", "pdf", "png")
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{cell_anno_var}{Character vector. Cell variables for which color bars are to be added to the heatmap. Default is NULL}

\item{clustMethod}{Character. Clustering method for ranking SNPs. Options are "hclust" and "skmean". Default is "hclust"}

\item{outputFormat}{Character. Output file type. Default is "" which outputs to the standard output}
}
\value{
A SNACSList object
}
\description{
Cluster cells, where the number of clusters is the number of hashes, using the mutation data. Then rank the SNPs in the order that best separates into those clusters. Generate heatmap of the mutation data with the ranked SNPs.
}
