% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hashMat}
\alias{hashMat}
\title{Example hash antibody data}
\format{
\subsection{\code{hashMat}}{

A matrix with 3 rows and 750 columns:
}
}
\usage{
hashMat
}
\description{
The antibody data should be a matrix with hashes in the rows and cells in the columns. The hash antibody value should be a continuous numeric measure.
...
}
\keyword{datasets}
