% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{imputeMissingMutations}
\alias{imputeMissingMutations}
\title{Imputes missing mutations in SNACSList object.}
\usage{
imputeMissingMutations(snacsObj, verbose = FALSE)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{verbose}{Logical. Prints information when running the method. Default is FALSE}
}
\value{
A SNACSList object
}
\description{
Imputes missing mutations in SNACSList object.
}
