% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSNACScall.R
\name{makeSnacsCall}
\alias{makeSnacsCall}
\title{Make SNACS calls}
\usage{
makeSnacsCall(
  snacsObj,
  minSampleSize = 100,
  cellProportionAboveBackgnd = 0.5,
  minClustSize = 2,
  clustComparePValue = 10^-5,
  maxClustSampleSize = Inf,
  clustCompareMethod = c("t", "hotelling"),
  maxClustSizeRnd2 = 100,
  dataTypeRnd2 = c("euclidean", "sum of squares", "log2 euclidean",
    "log2 sum of squares"),
  cbsAlpha = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{minSampleSize}{Integer. Minimum number of cells that is estimated by SNACS to be in a constituent sample. Gives a warning if below this number. Default is 100}

\item{cellProportionAboveBackgnd}{Numeric. Proportion of cells in a cluster that have to have signal above the background of a hash, determined by "backgndThreshold" parameter, for the cluster to be assigned to the hash. Default is 0.5. Range is 0-1}

\item{minClustSize}{Numeric. Minimum number of cells required to be in a cluster. Default is 2}

\item{clustComparePValue}{Numeric. P-value threshold to compare cluster pairs. Default is 10^-5}

\item{maxClustSampleSize}{Numeric. Maximum number of cells in a cluster that will be used to compare. If more, then this number of cells will be sampled. Default is Inf}

\item{clustCompareMethod}{Character. Test used to compare clusters. Default is t-test}

\item{maxClustSizeRnd2}{Integer. Maximum number of cells required to be in a cluster for making second round of SNACS calls. Default is 100}

\item{dataTypeRnd2}{Character. Type of data to be used for splitting the cells when making second round of SNACS calls. Default is euclidean distance of the cells to the cluster means from first round of SNACS calls}

\item{cbsAlpha}{Numeric. Significance level passed to DNAcopy::segment to for splitting the cells when making second round of SNACS calls. Default is 0.1}

\item{verbose}{Logical. Prints information when running the method. Default is FALSE}
}
\value{
A SNACSList object
}
\description{
The SNACS calls are made in two rounds. In round 1, the cells are clustered based on select SNPs. All cells in a cluster are assigned to a sample if > 50\% of cells from that cluster have a hash expression that exceeds the 95th percentile of the background distribution for that sample. Clusters assigned to multiple hashes are designated as multiplets. In round 2, the multiplet detection is refined by segmenting the antibody data and assigning a segment to a sample when it has > 50\% cells that exceeds the 75th percentile of the hash background. Only narrow segments (<=100 cells) are considered in this round. The results are stored in the columns "snacsRnd1" and "snacsRnd2" in the "annCell" table of the SNACS object
}
