% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutMat}
\alias{mutMat}
\title{Mutation data}
\format{
\subsection{\code{mutMat}}{

A matrix with 24 rows and 750 columns:
}
}
\usage{
mutMat
}
\description{
The mutation data should be a numeric matrix with SNPs in the rows and cells in the columns. 1s represent mutations and 0s non-mutations.
...
}
\keyword{datasets}
