% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{print.SNACSList}
\alias{print.SNACSList}
\title{Print SNACSList object.}
\usage{
\method{print}{SNACSList}(x, ...)
}
\arguments{
\item{x}{SNACSList object}

\item{...}{ignored}
}
\description{
Print SNACSList object.
}
