% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDoubletD.R
\name{runDoubletD}
\alias{runDoubletD}
\title{Run doubletD}
\usage{
runDoubletD(df_total, df_alt)
}
\arguments{
\item{df_total}{Data frame. First column is the cell ID. Rest of the columns repesent SNPs having total depth values}

\item{df_alt}{Data frame. First column is the cell ID. Rest of the columns repesent SNPs having alternate depth values}
}
\value{
A single column matrix of singlet or doublet designation for each cell
}
\description{
doubletD assigns singlet or doublet designation to each cell based on total depth and alternate depth values.
}
