% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSNACScall.R
\name{runSNACS}
\alias{runSNACS}
\title{Run SNACS to make demultiplex samples in a multiplexed SNP experiment}
\usage{
runSNACS(
  snacsObj,
  proportionMissingPerSNP = 0.4,
  proportionMissingPerCell = 0.4,
  proportionMutatedPerSNP = c(0.05, 0.95),
  proportionMutatedPerCell = c(0, 1),
  numSNP = 3,
  minSampleSize = 100,
  bgndThresDetMethod = c("automatic", "manual", "default modes", "two modes"),
  backgndThreshold = 0.95,
  backgndThresRnd2 = 0.75,
  cellProportionAboveBackgnd = 0.5,
  cellProportionBelowBackgndMode = 0.6,
  cellProportionForModeDetection = 0.75,
  hashThreshold = 0.5,
  bgndQuantileThreshold = NA,
  bgndQuantileThresRnd2 = NA,
  minClustSize = 2,
  clustComparePValue = 10^-5,
  maxClustSampleSize = Inf,
  clustCompareMethod = c("t", "hotelling"),
  maxClustSizeRnd2 = 100,
  dataTypeRnd2 = c("euclidean", "sum of squares", "log2 euclidean",
    "log2 sum of squares"),
  cbsAlpha = 0.1,
  outputFormat = c("", "pdf", "png", "none"),
  verbose = FALSE
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{proportionMissingPerSNP}{Numeric. Only SNPs with lower than this proportion of missing values are considered. Default is 0.4. Range is 0-1}

\item{proportionMissingPerCell}{Numeric. Only cells with lower than this proportion of missing values are considered. Default is 0.4. Range is 0-1}

\item{proportionMutatedPerSNP}{Numeric vector. Only SNPs having proportion of mutations within this range are considered. Default is c(0.05,0.95). Range is 0-1}

\item{proportionMutatedPerCell}{Numeric vector. Only cells having proportion of mutations within this range are considered. Default is c(0,1). Range is 0-1}

\item{numSNP}{Integer. Number of pairs of SNPs (high in one sample / low in second sample and vice versa) that best separate sample-pairs. There will be s times n(n-1) number of SNPs where s is the number of SNPs and n is the number of samples. Default is 3}

\item{minSampleSize}{Integer. Minimum number of cells that is estimated by SNACS to be in a constituent sample. Gives a warning if below this number. Default is 100}

\item{bgndThresDetMethod}{Character. Method for detecting threshold of background antibody distribution. Default is "automatic"}

\item{backgndThreshold}{Numeric. Threshold of the background antibody distribution of a hash above which the antibody will be considered to be expressed in a cell for making first round of SNACS calls. Default is 0.95. Range is 0-1}

\item{backgndThresRnd2}{Numeric. Threshold of the background antibody distribution of a hash above which the antibody will be considered to be expressed in a cell for making second round of SNACS calls. Default is 0.75. Range is 0-1}

\item{cellProportionAboveBackgnd}{Numeric. Proportion of cells in a cluster that have to have signal above the background of a hash, determined by "backgndThreshold" parameter, for the cluster to be assigned to the hash. Default is 0.5. Range is 0-1}

\item{cellProportionBelowBackgndMode}{Numeric. Maximum proportion of cells which can be below the mode of the estimated hash background distribution. Default is 0.6. Range is 0-1}

\item{cellProportionForModeDetection}{Numeric. Proportion of cells used to estimate mode of the background distribution. Used only if "cellProportionBelowBackgndMode" threshold is not met; otherwise, all cells are used. Default is 0.75. Range is 0-1}

\item{hashThreshold}{Numeric. Threshold of the hash value above which the antibody will be considered to be expressed in a cell. If NA then backgndThreshold is used. Default is 0.5}

\item{bgndQuantileThreshold}{Numeric. Threshold of the hash value above which the antibody will be considered to be expressed in a cell in first round of SNACS call. Default is NA. Used if bgndThresDetMethod = "manual"}

\item{bgndQuantileThresRnd2}{Numeric. Threshold of the hash value above which the antibody will be considered to be expressed in a cell in second round of SNACS call. Default is NA. Used if bgndThresDetMethod = "manual"}

\item{minClustSize}{Numeric. Minimum number of cells required to be in a cluster. Default is 2}

\item{clustComparePValue}{Numeric. P-value threshold to compare cluster pairs. Default is 10^-5}

\item{maxClustSampleSize}{Numeric. Maximum number of cells in a cluster that will be used to compare. If more, then this number of cells will be sampled. Default is Inf}

\item{clustCompareMethod}{Character. Test used to compare clusters. Default is t-test}

\item{maxClustSizeRnd2}{Integer. Maximum number of cells required to be in a cluster for making second round of SNACS calls. Default is 100}

\item{dataTypeRnd2}{Character. Type of data to be used for splitting the cells when making second round of SNACS calls. Default is euclidean distance of the cells to the cluster means from first round of SNACS calls}

\item{cbsAlpha}{Numeric. Significance level passed to DNAcopy::segment to for splitting the cells when making second round of SNACS calls. Default is 0.1}

\item{outputFormat}{Character. Output file type. "none" means nothing is plotted. Default is "" which outputs to the standard output}

\item{verbose}{Logical. Prints information when running the method. Default is FALSE}
}
\value{
A SNACSList object
}
\description{
Make calls for constituent samples in a multiplexed SNP experiment using SNP and hash antibody data.
}
