% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDoubletD.R
\name{runSNACSplusDoubletD}
\alias{runSNACSplusDoubletD}
\title{Add doubletD calls to SNACS calls}
\usage{
runSNACSplusDoubletD(snacsObj, verbose = FALSE)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{verbose}{Logical. Prints information when running the method. Default is FALSE}
}
\value{
A SNACSList object
}
\description{
First run doubletD, which assigns singlet or doublet designation to each cell based on total depth and alternate depth values. Then designate those cells as a doublet that are called so by doubletD but are called a singlet by SNACS Round 2. The result is stored in the column "snacsPlusDoubletD" in the "annCell" table of the SNACS object.
}
