% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{SNACSList}
\alias{SNACSList}
\title{Creates a SNACSList object.}
\usage{
SNACSList(
  mut,
  hashes,
  exptName,
  annCell = NULL,
  annSNP = NULL,
  annHash = NULL,
  hashColors = NULL
)
}
\arguments{
\item{mut}{Integer matrix. Matrix having mutation status as 0s and 1s. Rows depict SNPs and columns  cells}

\item{hashes}{Numeric matrix. Matrix having hash values as numeric data. Rows depict hashes and columns cells}

\item{exptName}{Character. Experiment name}

\item{annCell}{Data frame. Table of cell annotation. Default is NULL}

\item{annSNP}{Data frame. Table of SNP annotation. Default is NULL}

\item{annHash}{Data frame. Table of hash annotation. Default is NULL}

\item{hashColors}{Character vector. hashColors Colors of the hashes. Default is NULL}
}
\value{
A SNACSList object
}
\description{
SNACS stores data in a simple list-based data object called a SNACSList.
}
