% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSNACScall.R
\name{clusterSampleWithAntibodyData}
\alias{clusterSampleWithAntibodyData}
\title{Group cells based on hash antibody data}
\usage{
clusterSampleWithAntibodyData(
  snacsObj,
  minSampleSize = 100,
  bgndThresDetMethod = c("automatic", "manual", "default modes", "two modes"),
  backgndThreshold = 0.95,
  backgndThresRnd2 = 0.75,
  cellProportionBelowBackgndMode = 0.6,
  cellProportionForModeDetection = 0.75,
  hashThreshold = 0.5,
  bgndQuantileThreshold = NA,
  bgndQuantileThresRnd2 = NA,
  verbose = FALSE
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{minSampleSize}{Integer. Minimum number of cells that is estimated by SNACS to be in a constituent sample. Gives a warning if below this number. Default is 100}

\item{bgndThresDetMethod}{Character. Method for detecting threshold of background antibody distribution. Default is "automatic"}

\item{backgndThreshold}{Numeric. Threshold of the background antibody distribution of a hash above which the antibody will be considered to be expressed in a cell for making first round of SNACS calls. Default is 0.95. Range is 0-1}

\item{backgndThresRnd2}{Numeric. Threshold of the background antibody distribution of a hash above which the antibody will be considered to be expressed in a cell for making second round of SNACS calls. Default is 0.75. Range is 0-1}

\item{cellProportionBelowBackgndMode}{Numeric. Maximum proportion of cells which can be below the mode of the estimated hash background distribution. Default is 0.6. Range is 0-1}

\item{cellProportionForModeDetection}{Numeric. Proportion of cells used to estimate mode of the background distribution. Used only if "cellProportionBelowBackgndMode" threshold is not met; otherwise, all cells are used. Default is 0.75. Range is 0-1}

\item{hashThreshold}{Numeric. Threshold of the hash value above which the antibody will be considered to be expressed in a cell. If NA then backgndThreshold is used. Default is 0.5}

\item{bgndQuantileThreshold}{Numeric. Threshold of the hash value above which the antibody will be considered to be expressed in a cell in first round of SNACS call. Default is NA. Used if bgndThresDetMethod = "manual"}

\item{bgndQuantileThresRnd2}{Numeric. Threshold of the hash value above which the antibody will be considered to be expressed in a cell in second round of SNACS call. Default is NA. Used if bgndThresDetMethod = "manual"}

\item{verbose}{Logical. Prints information when running the method. Default is FALSE}
}
\description{
A cell is assigned to a specific sample if the antibody expression of that sample is expressed highly for that cell. These initial classifications are used to select SNPs for making SNACS calls.
}
