% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSNACScall.R
\name{clusterSampleWithAntibodyData}
\alias{clusterSampleWithAntibodyData}
\title{Assign sample IDs to cells based on hash antibody data}
\usage{
clusterSampleWithAntibodyData(
  snacsObj,
  backgndThreshold = 0.95,
  cellProportionBelowBackgndMode = 0.6,
  cellProportionForModeDetection = 0.75,
  hashThreshold = 0.5
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{backgndThreshold}{Numeric. Threshold of the background antibody distribution of a hash above which the antibody will be considered to be expressed in a cell. Default is 0.95. Range is 0-1}

\item{cellProportionBelowBackgndMode}{Numeric. Maximum proportion of cells which can be below the mode of the estimated hash background distribution. Default is 0.6. Range is 0-1}

\item{cellProportionForModeDetection}{Numeric. Proportion of cells used to estimate mode of the background distribution. Used only if "cellProportionBelowBackgndMode" threshold is not met; otherwise, all cells are used. Default is 0.75. Range is 0-1}

\item{hashThreshold}{Numeric. Threshold of the hash value above which the antibody will be considered to be expressed in a cell. Default is 0.5. Used if backgndThreshold = NA}
}
\description{
Assign sample IDs to cells based on hash antibody data
}
