% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHeatmap.R
\name{createHeatmap}
\alias{createHeatmap}
\title{Generate heatmap of mutation data}
\usage{
createHeatmap(
  snacsObj,
  cell_anno_var,
  cell_anno_name = NULL,
  col_dend = T,
  row_dend = F,
  h_title = NULL,
  outputFileName = "heatmap",
  outputFormat = c("", "pdf", "png", "none"),
  write2Table = F
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{cell_anno_var}{Character vector. Cell variables for which color bars are to be added to the heatmap. Default is NULL}

\item{cell_anno_name}{Character vector. Names of column color bars. Default is NULL}

\item{col_dend}{Logical. Displays cell dendrogram. Default is TRUE}

\item{row_dend}{Logical. Displays SNP dendrogram. Default is FALSE}

\item{h_title}{Character. Gives the heatmap output a title. Default is NULL}

\item{outputFileName}{Character. Output file name. Default is "heatmap"}

\item{outputFormat}{Character. Output file type. Default is "" which outputs to the standard output}

\item{write2Table}{Logical. Write the cell & SNP annotation tables ordered as in heatmap. Default is FALSE}
}
\value{
A SNACSList object
}
\description{
Generate heatmap of mutation data
}
