% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestSNPs.R
\name{getBestSNPs}
\alias{getBestSNPs}
\title{Select SNPs that best separate the hashes.}
\usage{
getBestSNPs(
  snacsObj,
  cell_anno_var = NULL,
  clustMethodForRankedSNPs = c("hclust", "skmean"),
  clustMethodForBestSNPs = c("hclust", "skmean"),
  pvSnpThres = 0,
  outputFormat = c("", "pdf", "png")
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{cell_anno_var}{Optional character vector. Cell variables for which color bars are to be added to the heatmap. Default is NULL}

\item{clustMethodForRankedSNPs}{Character. Clustering method for ranking SNPs. Options are "hclust" and "skmean". Default is "hclust"}

\item{clustMethodForBestSNPs}{Character. Clustering method for selecting best SNPs. Options are "hclust" and "skmean". Default is "hclust"}

\item{pvSnpThres}{P-value cutoff to select best SNPs}

\item{outputFormat}{Character. Output file type. Default is "" which outputs to the standard output}
}
\value{
A SNACSList object
}
\description{
Cluster data to rank and select SNPs that best separate the hashes. Generate heatmaps of the ranked and best selected SNPs.
}
