% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestSNPs.R
\name{getBestSNPs}
\alias{getBestSNPs}
\title{Select SNPs that best separate the hashes}
\usage{
getBestSNPs(
  snacsObj,
  cell_anno_var = NULL,
  clustMethodForRankedSNPs = c("hclust", "skmean"),
  clustMethodForBestSNPs = c("hclust", "skmean"),
  pvSnpThres = 10^-320,
  outputFormat = c("", "pdf", "png")
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{cell_anno_var}{Character vector. Cell variables for which color bars are to be added to the heatmap. Default is NULL}

\item{clustMethodForRankedSNPs}{Character. Clustering method for ranking SNPs. Options are "hclust" and "skmean". Default is "hclust"}

\item{clustMethodForBestSNPs}{Character. Clustering method for selecting best SNPs. Options are "hclust" and "skmean". Default is "hclust"}

\item{pvSnpThres}{Numeric. P-value cutoff to select best SNPs. Default is 10^-320}

\item{outputFormat}{Character. Output file type. Default is "" which outputs to the standard output}
}
\value{
A SNACSList object
}
\description{
Cluster cells, where the number of clusters is the number of hashes, using the mutation data. Rank the SNPs in the order that best separates into those clusters. Then select the best SNPs that separate the cells into those clusters and re-cluster the cells using those SNPs. Generate heatmaps of the mutation data with the ranked and best SNPs.
}
