% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSNACScall.R
\name{getHashModeInfo}
\alias{getHashModeInfo}
\title{Get mode information of density function}
\usage{
getHashModeInfo(mm, bw = "nrd0")
}
\arguments{
\item{mm}{.}

\item{bw}{.}
}
\value{
Mode and adjustment values
}
\description{
Get mode information of two mode density function of hash antibody data for background detection.
}
