% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSNACScall.R
\name{getModes}
\alias{getModes}
\title{Detect density function modes}
\usage{
getModes(x, bw, adjust, signifi, from, to)
}
\arguments{
\item{x}{Numeric. Hash antibody data}

\item{bw}{Character. Smoothing bandwidth parameter passed to density function}

\item{adjust}{Character. Bandwidth adjustment factor parameter passed to density function}

\item{signifi}{Numeric. Number of significant digits of the density modes, Used to order the modes}

\item{from}{Numeric. left-most point value passed to density function}

\item{to}{Numeric. right-most point value passed to density function}
}
\value{
A vector of modes
}
\description{
Find the modes of the density function of the hash antibody distribution for background detection.
}
