% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{h5readForSNACS}
\alias{h5readForSNACS}
\title{Read a HDF5 file and transform the data into a format that can be processed by SNACSList function. The datasets expected to be in the file are.
CELL_BARCODES - vector of cell barcodes
VARIANTS - vector of variant names
GT - genotype data matrix
ABS/clr - hash antibody data matrix
AB_DESCRIPTIONS - hash names
DP - total depth matrix
AD - alternate depth matrix
See h5read function in rhdf5 package and SNACSList function in SNACS packge for details}
\usage{
h5readForSNACS(file)
}
\arguments{
\item{file}{Character. Name of HDF5 file which has the data}
}
\value{
A list having objects that are needed to create a SNACS object
}
\description{
Read a HDF5 file and transform the data into a format that can be processed by SNACSList function. The datasets expected to be in the file are.
CELL_BARCODES - vector of cell barcodes
VARIANTS - vector of variant names
GT - genotype data matrix
ABS/clr - hash antibody data matrix
AB_DESCRIPTIONS - hash names
DP - total depth matrix
AD - alternate depth matrix
See h5read function in rhdf5 package and SNACSList function in SNACS packge for details
}
