% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{imputeMissingMutations}
\alias{imputeMissingMutations}
\title{Impute missing mutations in SNACSList object.}
\usage{
imputeMissingMutations(snacsObj, seed = 12345, verbose = FALSE)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{seed}{Integer. Not implemented. Seed for generating random number. Default is 12345}

\item{verbose}{Logical. Prints information when running the method. Default is FALSE}
}
\value{
A SNACSList object
}
\description{
Impute missing mutations in SNACSList object.
}
