% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashCall.R
\name{makeHashCall}
\alias{makeHashCall}
\title{Make hash calls.}
\usage{
makeHashCall(
  snacsObj,
  prob_ecdf = 0.95,
  probBackgndPeak = 0.6,
  probBelowForeground = 0.75,
  propAboveBackground = 0.5,
  minClustSize = 10,
  clustComparePValue = 10^-5
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{prob_ecdf}{.}

\item{probBackgndPeak}{.}

\item{probBelowForeground}{.}

\item{propAboveBackground}{.}

\item{minClustSize}{Integer. Minimum number of cells required to be in a cluster. Default is 10}

\item{clustComparePValue}{P-value cutoff to compare cluster pairs}
}
\value{
A SNACSList object
}
\description{
Make hash call based on mutation and hash data.
}
