% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashCall.R
\name{makeHashCall}
\alias{makeHashCall}
\title{Make hash calls.}
\usage{
makeHashCall(
  snacsObj,
  backgndThreshold = 0.95,
  cellProportionBelowBackgndMode = 0.6,
  cellProportionForModeDetection = 0.75,
  cellProportionAboveBackgnd = 0.5,
  minClustSize = 10,
  minClustSizeRnd2 = Inf,
  backgndThresRnd2 = 0.75,
  clustComparePValue = 10^-5,
  maxClustSampleSize = Inf,
  clustCompareMethod = c("t", "hotelling")
)
}
\arguments{
\item{snacsObj}{SNACSList object}

\item{backgndThreshold}{Threshold of the background antibody distribution of a hash above which the antibody will be considered to be expressed in a cell. Default is 0.95. Range is 0-1}

\item{cellProportionBelowBackgndMode}{Maximum proportion of cells which can be below the mode of the estimated hash background distribution. Default is 0.6. Range is 0-1}

\item{cellProportionForModeDetection}{Proportion of cells used to estimate mode of the background distribution. Used only if "cellProportionBelowBackgndMode" threshold is not met; otherwise, all cells are used. Default is 0.75. Range is 0-1}

\item{cellProportionAboveBackgnd}{Proportion of cells in a cluster that have to have signal above the background of a hash, determined by "backgndThreshold" parameter, for the cluster to be assigned to the hash. Default is 0.5. Range is 0-1}

\item{minClustSize}{Minimum number of cells required to be in a cluster. Default is 10}

\item{minClustSizeRnd2}{Minimum number of cells required to be in a cluster for making second round of hash calls. Inf means the second round is not run. Default is Inf}

\item{backgndThresRnd2}{Threshold of the background antibody distribution of a hash above which the antibody will be considered to be expressed in a cell for making second round of hash calls. Default is 0.75. Range is 0-1}

\item{clustComparePValue}{P-value threshold to compare cluster pairs}

\item{maxClustSampleSize}{Maximum number of cells in a cluster that will be used to compare. If more, then this number of cells will be sampled. Default is 5000}

\item{clustCompareMethod}{Test used to compare clusters. Default is t-test}
}
\value{
A SNACSList object
}
\description{
Make hash calls based on mutation and hash antibody data. The cluster information from the mutation data and the hash antibody data are used to make hash calls. Each cell is assigned one of the hash IDs or called a "Doublet" when it is a mixture of hashes. The distribution of the antibody measure of a hash is expected to be a mixture of a background and a foreground signal where the latter signals the presence of antibody for the hash. Expecting the mutation data of cells with similar hash profiles to cluster together, the cells are split into sub-clusters and each cluster is assigned a hash ID if enough proportion of cells in the cluster have signal above the background for that hash. The empirical background distribution of a hash antibody is estimated by taking the background mode of the data and generating an empirical symmetric distribution around it. The sub-clusters are created using the cell clusters from the second round of hierarchical clustering of the mutation data. First the antibody measures of the top two cluster pairs are compared using Hotelling's T2 test. If there is significant difference , the two clusters are splits into sub-clusters and each of those cluster pairs are tested for difference. The tree is traversed until there are no pairs of significantly different cluster pairs. The splitting of a cluster stops it reaches a minimum number of cells.
}
