% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDoubletD.R
\name{runDoubletD}
\alias{runDoubletD}
\title{Run doubletD}
\usage{
runDoubletD(snacsObj = NULL, depthTotal = NULL, depthAlt = NULL)
}
\arguments{
\item{snacsObj}{SNACSList object. Either snacsObj or depthTotal and depthAlt has to be provided}

\item{depthTotal}{Numeric matrix. Matrix having total depth values as numeric data. Rows depict SNPs and columns cells}

\item{depthAlt}{Numeric matrix. Matrix having total alternate values as numeric data. Rows depict SNPs and columns cells}
}
\value{
A SNACSList object if input is a SNACSList; otherwise, a vector of singlet or doublet designation for each cell
}
\description{
Assign singlet or doublet designation to each cell based on total depth and alternate depth values by applying doubletD method. The result is stored in the column "doubletD" in the "annCell" table of the SNACS object.
}
