% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDoubletD.R
\name{runSNACSplusDoubletD}
\alias{runSNACSplusDoubletD}
\title{Add doubletD calls to SNACS hash calls}
\usage{
runSNACSplusDoubletD(snacsObj)
}
\arguments{
\item{snacsObj}{SNACSList object}
}
\value{
A SNACSList object
}
\description{
First run doubletD, which assigns singlet or doublet designation to each cell based on total depth and alternate depth values. Then assign those cells called a doublet by doubletD but a singlet in the last round of SNACS hash calls made a doublet. The result is stored in the column "snacsPlusDoubletDCall" in the "annCell" table of the SNACS object.
}
