% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{SNACSList}
\alias{SNACSList}
\title{Creates a SNACSList object.}
\usage{
SNACSList(
  mut,
  hashes,
  exptName,
  annCell = NULL,
  annSNP = NULL,
  annHash = NULL,
  hashColors = NULL
)
}
\arguments{
\item{mut}{Matrix having mutation status as 0s and 1s}

\item{hashes}{Matrix having hash values as numeric data}

\item{exptName}{Experiment name}

\item{annCell}{Data frame of cell annotation}

\item{annSNP}{Data frame of SNP annotation}

\item{annHash}{Data frame of hash annotation}

\item{hashColors}{Colors of the hashes}
}
\value{
A SNACSList object
}
\description{
SNACS stores data in a simple list-based data object called a SNACSList.
}
