% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{SNACSList}
\alias{SNACSList}
\title{Create a SNACSList object.}
\usage{
SNACSList(
  mut,
  hashes,
  exptName,
  depthTotal = NULL,
  depthAlt = NULL,
  annCell = NULL,
  annSNP = NULL,
  annHash = NULL,
  hashColors = NULL
)
}
\arguments{
\item{mut}{Integer matrix. Matrix having mutation status as 0s and 1s. Rows depict SNPs and columns cells}

\item{hashes}{Numeric matrix. Matrix having hash values as numeric data. Rows depict hashes and columns cells}

\item{exptName}{Character. Experiment name}

\item{depthTotal}{Numeric matrix. Matrix having total depth values as numeric data. Rows depict SNPs and columns cells}

\item{depthAlt}{Numeric matrix. Matrix having total alternate values as numeric data. Rows depict SNPs and columns cells}

\item{annCell}{Data frame. Table of cell annotation. Default is NULL}

\item{annSNP}{Data frame. Table of SNP annotation. Default is NULL}

\item{annHash}{Data frame. Table of hash annotation. Default is NULL}

\item{hashColors}{Character vector. hashColors Colors of the hashes. Default is NULL}
}
\value{
A SNACSList object
}
\description{
Create a simple list-based data object called SNACSList that can be used by the methods in the SNACS package to make SNACS calls.
}
